General updates, mainly relating to hp-adaptivity. Also some bug fixes in parallel.
Modifications to the src directory:
-----------------------------------

 -- Changed how master and slave edges were determined for mortaring.
    Now we choose the longest edge as master, or the edge with the lower
    p-order as master if both edges have the same length. While the previous
    formulation was mathematically correct in its imposition of weak
    continuity, this version makes more sense and recovers strong continuity
    in more cases.

 -- Interface change: made PRefineableQElement<3,INITIAL_NNODE_1D>::oc_hang_helper()
    virtual.

 -- Generalised hp-refineable elements to 3D case.

 -- Removed superfluous functions
    interpolated_zeta_on_edge/face_before_p_refinement().

 -- Added 3D versions of p-refineable Navier-Stokes elements. Also changed 2D
    PRefineableQCrouzeixRaviartElements so that for NNODE_1D=3 linear rather
    than piecewise constant pressure interpolation is used. This is the same
    as the standard QCrouzeixRaviartElements.

 -- Made NavierStokesEquations::interpolated_p_nst() virtual so that it can be
    overridden to calculate enriched pressures.

 -- Added p-refineable versions of linear elasticity elements

 -- Fixed problem with halo nodes' positions differing from their haloed
    counterparts in situations where the node was created on the haloed
    process by an element that does not exist on the halo processor.
    e.g. in the case of h-refinement of elements with different p-orders
    that share an edge on the outer edge of the halo layer on one of the
    processors.
    This requires communication of nodal positions from the haloed nodes to
    the corresponding halo nodes. If the stored global coordinate of the
    haloed node is not consistent with its interpolated coordinate then this
    is sent to corresponding nodes that may be shared with elements that do
    not exist on the halo processor and were therefore created with a
    different global coordinate. This is done in the new function
    TreeBasedRefineableMeshBase::synchronise_nonhanging_nodes(), called just
    before synchronise_hanging_nodes().

 -- Updated the missing master node functions to search in more places for
    missing nodes before giving up.

 -- Updated the missing master node functions to handle MacroElementNodeUpdate
    stuff. Now fully functioning MacroElementNodeUpdateElements can be created
    for missing master nodes that get added to the external storage in order
    to facilitate their node updates.

     * This required an INTERFACE CHANGE to the function initial_setup()
       that now takes an optional argument const unsigned &initial_p_order=0
       that allows the initial p-order to be specified for elements not in a
       tree.

    However, this functionality causes some problems for specific processor
    layouts when "duplicate" nodes are deleted. So instead an error is thrown
    inside additional_synchronise_hanging_nodes() if
    MacroElementNodeUpdateNodes are used, are before any additional nodes are
    created.

 -- Added flag in RefineableMeshBase() to control whether additional
    synchronisation of hanging nodes is performed.

    The function TreeBasedRefineableMesh<ELEMENT>::additional_synchronise_hanging_nodes()
    is used when the masters of hanging nodes cannot be found on the local
    processor. If all the master nodes are located in
    synchronise_hanging_nodes() then the function is never called, as is the
    case when uniformly-spaced nodes are used. If some masters are not found
    then the required nodes are constructed on the local processor in the
    external storage and the corresponding nodes become haloed. This process
    requires communication. However, if this synchronisation is not required
    then the user may set the flag
    RefineableMeshBase::Additional_synchronisation_of_hanging_nodes_not_required=true
    to prevent this additional cost. The result will be the possibility that the
    hanging status of nodes on the outer edge of the halo layer could differ if
    nonconforming meshes are used with elements with nonuniformly-spaced nodes.

 -- New function TreeBasedRefineableMeshBase::synchronise_nonhanging_nodes().
    For any haloed node whose position is determined by an element that does not
    exist on the processor on which the halo copy exists, its position is
    communicated. This ensures that the nodal positions are consistent across
    all processors when using nonconforming meshes. Communication always occurs,
    but the ammount of data communicated is limited to just the positions of the
    nodes whose positions differ. For elements with uniformly spaced nodes no
    such nodes exist, and so each processor sends a single integer (0) to every
    other processor.

 -- Removed call to delete_all_external_storage() that caused hangs in
    refine_base_mesh().

 -- Removed MPI_Barrier() that caused hangs in Mesh::check_halo_schemes().

 -- Replaced some Comm_pt!=0 with mesh_is_distributed().

 -- Changed order of rebuild_from_sons() in
    MacroElementNodeUpdateElement<ELEMENT>

 -- Added alternative constructor for TElement to allow bypass of check for
    maximum NNODE_1D for higher orders.

 -- Removed BENFLAGS from nearly everywhere.

Modifications to the demo_drivers directory:
--------------------------------------------

 -- Added demo driver eighth_sphere_poisson_hp_adapt.cc to test 3D
    hp-adaptivity.

 -- Changed validata in demo drivers that use the updated mortaring algorithm.

Modifications to the doc directory:
-----------------------------------

 -- Changed equation markup in
    doc/linear_elasticity/periodic_load/periodic_load.txt so that it compiles
    with doxygen.



git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@952 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ src/generic/Qelements.h
@@ -937,7 +937,7 @@ public:
   {
    s.resize(2);
    unsigned j0=j%NNODE_1D;
-   unsigned j1=unsigned(double(j)/double(NNODE_1D));
+   unsigned j1=j/NNODE_1D;
    const double S_min = this->s_min();
    const double S_range = this->s_max() - S_min;
    s[0]=S_min+double(j0)/double(NNODE_1D-1)*S_range;
@@ -950,7 +950,7 @@ public:
   {
    s_fraction.resize(2);
    unsigned j0 = j%NNODE_1D;
-   unsigned j1 = unsigned(double(j)/double(NNODE_1D));
+   unsigned j1 = j/NNODE_1D;
    s_fraction[0] = double(j0)/double(NNODE_1D-1);
    s_fraction[1] = double(j1)/double(NNODE_1D-1);
   }
@@ -1302,8 +1302,8 @@ public:
   {
    s.resize(3);
    unsigned j0=j%NNODE_1D;
-   unsigned j1=unsigned(double(j)/double(NNODE_1D))%NNODE_1D;
-   unsigned j2=unsigned(double(j)/double(NNODE_1D*NNODE_1D));
+   unsigned j1=(j/NNODE_1D)%NNODE_1D;
+   unsigned j2=j/(NNODE_1D*NNODE_1D);
    const double S_min = this->s_min();
    const double S_range = this->s_max() - S_min;
 
@@ -1318,8 +1318,8 @@ public:
   {
    s_fraction.resize(3);
    unsigned j0 = j%NNODE_1D;
-   unsigned j1=unsigned(double(j)/double(NNODE_1D))%NNODE_1D;
-   unsigned j2=unsigned(double(j)/double(NNODE_1D*NNODE_1D));
+   unsigned j1=(j/NNODE_1D)%NNODE_1D;
+   unsigned j2=j/(NNODE_1D*NNODE_1D);
    s_fraction[0]= double(j0)/double(NNODE_1D-1);
    s_fraction[1]= double(j1)/double(NNODE_1D-1);
    s_fraction[2]= double(j2)/double(NNODE_1D-1);
